fix-polygon-behaviour-64
Fixed issue 'Strange polygon behaviour on 64-bit machines'
reported by Chris Mears at 14 Dec 2006.

Problem: the arrays inspected by the wxDC_Polygon function
are assumed to contain 32-bits ints. On 64-bit machines
these are actually 64-bits ints.

Fix: (a bit of a hack...) the code now assumes that the
arrays contain intptr_t values. This type represents ints
of the same number of bits as the Haskell ints (on 32-bit
and 64-bit machines). The code now dereferences each element
in the array correctly, but still truncates the possibly
64-bit value to a 32-bit value. This is not a problem
in practice.

darcs-hash:20070202154517-fafd9-62ab5a6bdbb39071c10fa27907085e6aac842321.gz

+++ wxc/src/ewxw/eljdc.cpp
@@ -76,7 +76,7 @@ EWXWEXPORT(void, wxDC_DrawPolygon)(void* _obj, int n, void* x, void* y, int xoff
 	wxPoint* lst = (wxPoint*)malloc (n * sizeof(wxPoint));
 	
 	for (int i = 0; i < n; i++)
-		lst[i] = wxPoint(((int*)x)[i], ((int*)y)[i]);
+		lst[i] = wxPoint(((intptr_t*)x)[i], ((intptr_t*)y)[i]);
 	
 	((wxDC*)_obj)->DrawPolygon(n, lst, (wxCoord)xoffset, (wxCoord)yoffset, fillStyle);
